Switch Vagrant nodes to use the latest Krates Worker official  image published

+++ build/deploy_gem.sh
@@ -2,14 +2,17 @@
 
 set -ue
 
-# login
-echo :rubygems_api_key: $RUBYGEMS_KEY > ~/.gem/credentials
-chmod 0600 ~/.gem/credentials
-
 # install dependencies
 bundle install
 gem build *.gemspec
-gem push *.gem
 
-# cleanup
-rm ~/.gem/credentials
+if [ -n "$RUBYGEMS_KEY" ]; then
+  # login
+  echo :rubygems_api_key: $RUBYGEMS_KEY > ~/.gem/credentials
+  chmod 0600 ~/.gem/credentials
+
+  gem push *.gem
+
+  # cleanup
+  rm ~/.gem/credentials
+fi
